trigger_function/requirements.txt
pydantic
requests
python-dotenv

google-cloud-firestore
google-cloud-pubsub
firebase-admin

functions-framework
flask